+++ pages/componentsB/datetime.vue
@@ -30,15 +30,21 @@
 			</u-form>
 		</view>
 		
-		<view class="title">页面查询条件</view>
-		<view class="padding">
+		<view class="title">自定义展示部分 (页面查询条件)</view>
+		<view class="section">
 			<hf-form-datetime v-model="model.startTime" mode="date">
-			    <template v-slot:display-section="{ valueName }">
-					<text>{{ valueName || '开始时间' }}</text>
-					<hf-icon name="arrow-down-fill" size="10"></hf-icon>
+				<template v-slot:display-section="{ valueName, visible }">
+					<u-icon
+						:name="visible ? 'arrow-up-fill' : 'arrow-down-fill'"
+						size="14"
+						:label="valueName || '开始时间'"
+						labelPos="left"
+					></u-icon>
 				</template>
 			</hf-form-datetime>
 		</view>
+		
+		<temp-data :info="model"></temp-data>
 	</view>
 </template>
 
@@ -47,7 +53,10 @@
 		data() {
 			return {
 				model: {
+					datetime: '',
+					date: '',
 					time: '12:24',
+					yearMonth: '',
 					startTime: ''
 				}
 			}
@@ -57,8 +66,4 @@
 
 <style lang="scss" scoped>
 	@import '@/common/styles/form.scss';
-	
-	.padding {
-		padding: 0 $df;
-	}
 </style>